[FEATURE-13] Move appStore to root folder
